More interface unification and cleanup.
Working on better validation of input and making it clear to the user what is happening as they are inputing data rather than throw an error after the fact which can be very frustrating when a person has wated their time doing something incorrectly. Better let them know before it becomes a problem that must be corrected later.